requirements.txt
gradio
numpy
opencv-python-headless
pillow
